Create initial DRAFT of MoOde guidelines.
Used guidelines from grrr-amsterdam as a starting point.
Most of the focus was on git & branching.
That should be a close starting point to what we should use,
based on discussion and decisions from the forums.

This DRAFT check-in provides framework for other efforts:
coding-standards, release process, developer environment, etc.
I have no really hard opinions there, just thought that this
could caputure a starting point, and a place to build upon.